fix a couple of casting issues with userPriority. Other than CalculationModifier this is an actual bug

+++ code/src/java/plugin/modifier/set/AddModifierFactory.java
@@ -179,7 +179,7 @@ public class AddModifierFactory<T> implements ModifierFactory<T[]>
 		@Override
 		public long getPriority()
 		{
-			return (userPriority << 32) + 3;
+			return ((long)userPriority << 32) + 3;
 		}
 
 		@Override
